pyproject.toml
[project]
name = "gradio-on-lambda-bedrock-aws"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Osman Orka", email = "osmanorka@example.com" }
]
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.96",
    "botocore[crt]>=1.42.96",
    "gradio>=6.13.0",
    "litellm>=1.83.13",
]

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "aws-cdk-lib>=2.251.0",
    "constructs>=10.6.0",
]
